# Optimize performance of `initcap` function (~2x faster)