--- a/datafusion/functions/Cargo.toml
+++ b/datafusion/functions/Cargo.toml
@@ -207,3 +207,8 @@ required-features = ["unicode_expressions"]
 harness = false
 name = "trunc"
 required-features = ["math_expressions"]
+
+[[bench]]
+harness = false
+name = "initcap"
+required-features = ["string_expressions"]