@tlm365 2024-12-08
I'm not sure if this is an intended implementation or not. But seems like it would be better if we return `StringViewArray` instead of `StringArray` [here](https://github.com/apache/datafusion/blob/6c9355d5be8b6045865fed67cb6d028b2dfc2e06/datafusion/functions/src/string/initcap.rs#L130) 🤔.